requirements.txt
streamlit==1.39.0
sqlmodel==0.0.22
sqlalchemy==2.0.36
python-dotenv==1.0.1
